Migrate providers from DisplayCategoryList to DisplayCategoryListWrap
