+++ src/pages/DisplayCategoryList/DisplayCategoryList.jsx
@@ -7,11 +7,6 @@ import {
   CategoriesProvider,
   VisibleComponent,
   usePageScriptsActionContext,
-  CartProvider,
-  ProductOptionProvider,
-  OrderSheetProvider,
-  TabsProvider,
-  useProductInquiryActionContext,
 } from '@shopby/react-components';
 
 import useLayoutChanger from '../../hooks/useLayoutChanger';
@@ -22,18 +17,11 @@ import CategoryMenu from './menu/CategoryMenu';
 const DisplayCategoryList = () => {
   const { t } = useTranslation('title');
 
-  const initialTabs = [
-    { value: 'DETAIL', label: t('상세정보') },
-    // Add more tabs as needed
-  ];
-
   const [searchParams] = useSearchParams();
   const keyword = searchParams.get('keyword') ?? '';
   const categoryNo = Number(searchParams.get('categoryNo'));
   const depth = Number(searchParams.get('depth') ?? 1);
   const { delayPageScriptLoading } = usePageScriptsActionContext();
-  const productNo = Number(searchParams.get('productNo'));
-  const { searchInquiries } = useProductInquiryActionContext();
 
   if (keyword) {
     useLayoutChanger({
@@ -56,50 +44,18 @@ const DisplayCategoryList = () => {
     delayPageScriptLoading();
   }, []);
 
-  useEffect(() => {
-    searchInquiries();
-
-    if (productNo > 0) {
-      fetchProductDetail({
-        productNo,
-        channelType,
-      });
-
-      fetchRelatedProducts({
-        productNo,
-      });
-
-      fetchSelectorOptions({
-        productNo,
-      });
-    }
-  }, [productNo]);
-
   return (
-    <TabsProvider
-      initialState={{
-        currentTab: 'DETAIL',
-        tabs: initialTabs,
-      }}
-    >
-      <OrderSheetProvider>
-        <CartProvider>
-          <ProductOptionProvider productNo={productNo}>
-            <ProductSearchProvider>
-              <VisibleComponent
-                shows={!keyword}
-                TruthyComponent={
-                  <CategoriesProvider>
-                    <CategoryMenu categoryNo={categoryNo} depth={depth} />
-                  </CategoriesProvider>
-                }
-              />
-              <DisplayCategoryListWrap />
-            </ProductSearchProvider>
-          </ProductOptionProvider>
-        </CartProvider>
-      </OrderSheetProvider>
-    </TabsProvider>
+    <ProductSearchProvider>
+      <VisibleComponent
+        shows={!keyword}
+        TruthyComponent={
+          <CategoriesProvider>
+            <CategoryMenu categoryNo={categoryNo} depth={depth} />
+          </CategoriesProvider>
+        }
+      />
+      <DisplayCategoryListWrap />
+    </ProductSearchProvider>
   );
 };
